refactor(core): home wordpiece unit tests
Move the WordPiece tokenizer tests next to the impl
(wordpiece/{mod.rs,tests.rs}). Public paths are unchanged.
